@@ -2,20 +2,50 @@
     "private": true,
     "scripts": {
         "dev": "npm run development",
-        "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "watch": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --watch --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "watch-poll": "npm run watch -- --watch-poll",
-        "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
+        "development": "mix",
+        "watch": "mix watch",
+        "watch-poll": "mix watch -- --watch-options-poll=1000",
+        "hot": "mix watch --hot",
         "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
+        "production": "mix --production"
     },
     "devDependencies": {
-        "axios": "^0.17",
-        "bootstrap-sass": "^3.3.7",
-        "cross-env": "^5.1",
-        "jquery": "^3.2",
-        "laravel-mix": "^1.0",
-        "lodash": "^4.17.4",
-        "vue": "^2.5.7"
+        "autoprefixer": "^10.4.13",
+        "axios": "^0.21",
+        "browser-sync": "^2.27.11",
+        "browser-sync-webpack-plugin": "^2.3.0",
+        "i": "^0.3.7",
+        "laravel-mix": "^6.0.6",
+        "laravel-vite-plugin": "^0.7.3",
+        "lodash": "^4.17.19",
+        "npm": "^9.4.1",
+        "postcss": "^8.1.14",
+        "resolve-url-loader": "^4.0.0",
+        "sass": "^1.49.7",
+        "sass-loader": "^12.4.0",
+        "vue-html-loader": "^1.2.4",
+        "vue-loader": "^16.8.3",
+        "vue-resource": "^1.5.3",
+        "vue-template-compiler": "^2.7.14",
+        "webpack": "^5.75.0"
+    },
+    "dependencies": {
+        "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+        "@headlessui/vue": "^1.7.9",
+        "@inertiajs/inertia": "^0.11.1",
+        "@inertiajs/inertia-vue3": "^0.6.0",
+        "@inertiajs/progress": "^0.2.7",
+        "@inertiajs/vue3": "^1.0.0",
+        "@popperjs/core": "^2.11.6",
+        "bootstrap": "^5.2.3",
+        "bootstrap-icons": "^1.10.3",
+        "human-date": "^1.4.0",
+        "inertia-title": "^1.2.0",
+        "jquery": "^3.6.3",
+        "pinia": "^2.0.30",
+        "popper.js": "^1.16.1",
+        "vue": "^3.2.45",
+        "vue-masonry": "^0.16.0",
+        "vue-social-sharing": "^4.0.0-alpha4"
     }
 }
