@@ -1,8 +1,9 @@
 <?php
 
-namespace App\Http;
+namespace Vanguard\Http;
 
 use Illuminate\Foundation\Http\Kernel as HttpKernel;
+use Vanguard\Http\Middleware\HandleInertiaRequests;
 
 class Kernel extends HttpKernel
 {
@@ -14,15 +15,14 @@ class Kernel extends HttpKernel
      * @var array
      */
     protected $middleware = [
+        \Vanguard\Http\Middleware\TrustProxies::class,
+        \Fruitcake\Cors\HandleCors::class,
+//        \Vanguard\Http\Middleware\VerifyInstallation::class,
         \Illuminate\Foundation\Http\Middleware\CheckForMaintenanceMode::class,
         \Illuminate\Foundation\Http\Middleware\ValidatePostSize::class,
-        \App\Http\Middleware\TrimStrings::class,
+        \Vanguard\Http\Middleware\TrimStrings::class,
         \Illuminate\Foundation\Http\Middleware\ConvertEmptyStringsToNull::class,
-        \App\Http\Middleware\TrustProxies::class,
-        \App\Http\Middleware\Language::class,
-        \App\Http\Middleware\Cors::class
     ];
-
     /**
      * The application's route middleware groups.
      *
@@ -30,35 +30,65 @@ class Kernel extends HttpKernel
      */
     protected $middlewareGroups = [
         'web' => [
-            \App\Http\Middleware\EncryptCookies::class,
+            \Vanguard\Http\Middleware\EncryptCookies::class,
             \Illuminate\Cookie\Middleware\AddQueuedCookiesToResponse::class,
             \Illuminate\Session\Middleware\StartSession::class,
             // \Illuminate\Session\Middleware\AuthenticateSession::class,
             \Illuminate\View\Middleware\ShareErrorsFromSession::class,
-            \App\Http\Middleware\VerifyCsrfToken::class,
+            \Vanguard\Http\Middleware\VerifyCsrfToken::class,
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
-            
+            HandleInertiaRequests::class,
+            'banned',
         ],
-
         'api' => [
+            \Vanguard\Http\Middleware\UseApiGuard::class,
+            \Laravel\Sanctum\Http\Middleware\EnsureFrontendRequestsAreStateful::class,
             'throttle:60,1',
             'bindings',
+            'banned',
         ],
     ];
 
     /**
      * The application's route middleware.
      *
-     * These middleware may be assigned to groups or used individually.
-     *
      * @var array
      */
     protected $routeMiddleware = [
-        'auth' => \Illuminate\Auth\Middleware\Authenticate::class,
+        'jwt' => \Vanguard\Http\Middleware\JWTMiddleware::class,
+        'auth' => \Vanguard\Http\Middleware\Authenticate::class,
         'auth.basic' => \Illuminate\Auth\Middleware\AuthenticateWithBasicAuth::class,
         'bindings' => \Illuminate\Routing\Middleware\SubstituteBindings::class,
+        'cache.headers' => \Illuminate\Http\Middleware\SetCacheHeaders::class,
         'can' => \Illuminate\Auth\Middleware\Authorize::class,
-        'guest' => \App\Http\Middleware\RedirectIfAuthenticated::class,
+        'guest' => \Vanguard\Http\Middleware\RedirectIfAuthenticated::class,
+        'signed' => \Illuminate\Routing\Middleware\ValidateSignature::class,
         'throttle' => \Illuminate\Routing\Middleware\ThrottleRequests::class,
+        'verified' => \Illuminate\Auth\Middleware\EnsureEmailIsVerified::class,
+        'registration' => \Vanguard\Http\Middleware\RegistrationEnabled::class,
+        'social.login' => \Vanguard\Http\Middleware\SocialLogin::class,
+        'role' => \Vanguard\Http\Middleware\CheckRole::class,
+        'permission' => \Vanguard\Http\Middleware\CheckPermissions::class,
+        'session.database' => \Vanguard\Http\Middleware\DatabaseSession::class,
+        'two-factor' => \Vanguard\Http\Middleware\TwoFactorEnabled::class,
+        'verify-2fa-phone' => \Vanguard\Http\Middleware\VerifyTwoFactorPhone::class,
+        'password-reset' => \Vanguard\Http\Middleware\PasswordResetEnabled::class,
+        'banned' => \Vanguard\Http\Middleware\CheckIfBanned::class,
+    ];
+
+    /**
+     * The priority-sorted list of middleware.
+     *
+     * This forces non-global middleware to always be in the given order.
+     *
+     * @var array
+     */
+    protected $middlewarePriority = [
+        \Illuminate\Session\Middleware\StartSession::class,
+        \Illuminate\View\Middleware\ShareErrorsFromSession::class,
+        \Vanguard\Http\Middleware\Authenticate::class,
+        \Illuminate\Session\Middleware\AuthenticateSession::class,
+        \Illuminate\Routing\Middleware\SubstituteBindings::class,
+        \Illuminate\Auth\Middleware\Authorize::class,
     ];
 }
