@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Providers;
+namespace Vanguard\Providers;
 
-use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Broadcast;
+use Illuminate\Support\ServiceProvider;
 
 class BroadcastServiceProvider extends ServiceProvider
 {
@@ -16,6 +16,21 @@ class BroadcastServiceProvider extends ServiceProvider
     {
         Broadcast::routes();
 
-        require base_path('routes/channels.php');
+        /*
+         * Authenticate the user's personal channel...
+         */
+        Broadcast::channel('Vanguard.User.*', function ($user, $userId) {
+            return (int)$user->id === (int)$userId;
+        });
+    }
+
+    /**
+     * Register the service provider.
+     *
+     * @return void
+     */
+    public function register()
+    {
+        //
     }
 }
