@@ -1,9 +1,17 @@
 <?php
 
-namespace App\Providers;
+namespace Vanguard\Providers;
 
-use Illuminate\Support\Facades\Event;
+use Illuminate\Auth\Events\Registered;
+use Illuminate\Auth\Events\Verified;
+use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
+use Vanguard\Events\User\Banned;
+use Vanguard\Events\User\LoggedIn;
+use Vanguard\Listeners\Login\UpdateLastLoginTimestamp;
+use Vanguard\Listeners\Registration\SendSignUpNotification;
+use Vanguard\Listeners\Users\ActivateUser;
+use Vanguard\Listeners\Users\InvalidateSessions;
 
 class EventServiceProvider extends ServiceProvider
 {
@@ -13,13 +21,32 @@ class EventServiceProvider extends ServiceProvider
      * @var array
      */
     protected $listen = [
-        'App\Events\Event' => [
-            'App\Listeners\EventListener',
+        Registered::class => [
+            SendEmailVerificationNotification::class,
+            SendSignUpNotification::class,
         ],
+        LoggedIn::class => [
+            UpdateLastLoginTimestamp::class
+        ],
+        Banned::class => [
+            InvalidateSessions::class
+        ],
+        Verified::class => [
+            ActivateUser::class
+        ]
     ];
 
     /**
-     * Register any events for your application.
+     * The subscriber classes to register.
+     *
+     * @var array
+     */
+    protected $subscribe = [
+        //
+    ];
+
+    /**
+     * Register any other events for your application.
      *
      * @return void
      */
