@@ -1,20 +1,28 @@
 <?php
 
-namespace App\Providers;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Route;
+namespace Vanguard\Providers;
+
 use Illuminate\Foundation\Support\Providers\RouteServiceProvider as ServiceProvider;
+use Route;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Vanguard\Repositories\Role\RoleRepository;
+use Vanguard\Repositories\Session\SessionRepository;
+use Vanguard\Repositories\User\UserRepository;
 
 class RouteServiceProvider extends ServiceProvider
 {
     /**
-     * This namespace is applied to your controller routes.
-     *
+     * This namespace is applied to the controller routes in your web routes file.
      * In addition, it is set as the URL generator's root namespace.
-     *
      * @var string
      */
-    protected $namespace = 'App\Http\Controllers';
+    protected $webNamespace = 'Vanguard\Http\Controllers\Web';
+
+    /**
+     * This namespace is applied to the controller routes in your api routes file.
+     * @var string
+     */
+    protected $apiNamespace = 'Vanguard\Http\Controllers\Api';
 
     /**
      * Define your route model bindings, pattern filters, etc.
@@ -23,9 +31,37 @@ class RouteServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        //
-
         parent::boot();
+
+        $this->bindUser();
+        $this->bindRole();
+        $this->bindSession();
+    }
+
+    private function bindUser()
+    {
+        $this->bindUsingRepository('user', UserRepository::class);
+    }
+
+    private function bindUsingRepository($entity, $repository, $method = 'find')
+    {
+        Route::bind($entity, function ($id) use ($repository, $method) {
+            if ($object = app($repository)->$method($id)) {
+                return $object;
+            }
+
+            throw new NotFoundHttpException("Resource not found.");
+        });
+    }
+
+    private function bindRole()
+    {
+        $this->bindUsingRepository('role', RoleRepository::class);
+    }
+
+    private function bindSession()
+    {
+        $this->bindUsingRepository('session', SessionRepository::class);
     }
 
     /**
@@ -33,45 +69,13 @@ class RouteServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    public function map(Request $request)
+    public function map()
     {
-        $this->mapApiRoutes();
-
-        $this->mapWebRoutes($request);
-
-        //
-    }
-
-    /**
-     * Define the "web" routes for the application.
-     *
-     * These routes all receive session state, CSRF protection, etc.
-     *
-     * @return void
-     */
-    protected function mapWebRoutes($request)
-    {
-
-        if(in_array($request->segment(1), config('app.available_locales')))
-        {
-            $locale = $request->segment(1);
-            //session('lang',$locale);
+        if ($this->app['config']->get('auth.expose_api')) {
+            $this->mapApiRoutes();
         }
-        else
-        {
-            $locale = null;
-        }
-        Route::group([
-            'middleware' => 'web',
-            'namespace' => $this->namespace,
-            'prefix' => $locale
-        ], function ($router) {
-            require base_path('routes/web.php');
-        });
-        /*
-        Route::middleware('web')
-             ->namespace($this->namespace)
-             ->group(base_path('routes/web.php'));)*/
+
+        $this->mapWebRoutes();
     }
 
     /**
@@ -83,9 +87,29 @@ class RouteServiceProvider extends ServiceProvider
      */
     protected function mapApiRoutes()
     {
-        Route::prefix('api')
-             ->middleware('api')
-             ->namespace($this->namespace)
-             ->group(base_path('routes/api.php'));
+        Route::group([
+            'middleware' => 'api',
+            'namespace' => $this->apiNamespace,
+            'prefix' => 'api',
+        ], function () {
+            require base_path('routes/api.php');
+        });
+    }
+
+    /**
+     * Define the "web" routes for the application.
+     *
+     * These routes all receive session state, CSRF protection, etc.
+     *
+     * @return void
+     */
+    protected function mapWebRoutes()
+    {
+        Route::group([
+            'namespace' => $this->webNamespace,
+            'middleware' => 'web',
+        ], function ($router) {
+            require base_path('routes/web.php');
+        });
     }
 }
