@@ -1,30 +1,134 @@
 <?php
 
-namespace App;
+namespace Vanguard;
 
-use Illuminate\Notifications\Notifiable;
+use Illuminate\Auth\Passwords\CanResetPassword;
+use Illuminate\Contracts\Auth\MustVerifyEmail;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
+use Illuminate\Notifications\Notifiable;
+use Laravel\Sanctum\HasApiTokens;
+use Mail;
+use Vanguard\Events\User\RequestedPasswordResetEmail;
+use Vanguard\Presenters\Traits\Presentable;
+use Vanguard\Presenters\UserPresenter;
+use Vanguard\Services\Auth\TwoFactor\Authenticatable as TwoFactorAuthenticatable;
+use Vanguard\Services\Auth\TwoFactor\Contracts\Authenticatable as TwoFactorAuthenticatableContract;
+use Vanguard\Support\Authorization\AuthorizationUserTrait;
+use Vanguard\Support\CanImpersonateUsers;
+use Vanguard\Support\Enum\UserStatus;
 
-class User extends \TCG\Voyager\Models\User
+class User extends Authenticatable implements TwoFactorAuthenticatableContract, MustVerifyEmail
 {
-    use Notifiable;
+    use TwoFactorAuthenticatable,
+        CanResetPassword,
+        Presentable,
+        AuthorizationUserTrait,
+        Notifiable,
+        CanImpersonateUsers,
+        HasApiTokens,
+        HasFactory;
+
+    protected $presenter = UserPresenter::class;
+
+    /**
+     * The database table used by the model.
+     *
+     * @var string
+     */
+    protected $table = 'users';
+
+    protected $dates = ['last_login', 'birthday'];
 
     /**
      * The attributes that are mass assignable.
      *
      * @var array
      */
-    
     protected $fillable = [
-        'name', 'email', 'password',
+        'email', 'password', 'username', 'first_name', 'last_name', 'phone', 'avatar',
+        'address', 'country_id', 'birthday', 'last_login', 'confirmation_token', 'status',
+        'remember_token', 'role_id', 'email_verified_at'
     ];
 
     /**
-     * The attributes that should be hidden for arrays.
+     * The attributes excluded from the model's JSON form.
      *
      * @var array
      */
-    protected $hidden = [
-        'password', 'remember_token',
-    ];
+    protected $hidden = ['password', 'remember_token'];
+
+    /**
+     * Always encrypt password when it is updated.
+     *
+     * @param $value
+     * @return string
+     */
+    public function setPasswordAttribute($value)
+    {
+        $this->attributes['password'] = bcrypt($value);
+    }
+
+    public function setBirthdayAttribute($value)
+    {
+        $this->attributes['birthday'] = trim($value) ?: null;
+    }
+
+    public function gravatar()
+    {
+        $hash = hash('md5', strtolower(trim($this->attributes['email'])));
+
+        return sprintf("https://www.gravatar.com/avatar/%s?size=150", $hash);
+    }
+
+    public function isUnconfirmed()
+    {
+        return $this->status == UserStatus::UNCONFIRMED;
+    }
+
+    public function isActive()
+    {
+        return $this->status == UserStatus::ACTIVE;
+    }
+
+    public function isBanned()
+    {
+        return $this->status == UserStatus::BANNED;
+    }
+
+    public function country()
+    {
+        return $this->belongsTo(Country::class, 'country_id');
+    }
+
+    /**
+     * Send the password reset notification.
+     *
+     * @param string $token
+     * @return void
+     */
+    public function sendPasswordResetNotification($token)
+    {
+        Mail::to($this)->send(new \Vanguard\Mail\ResetPassword($token));
+
+        event(new RequestedPasswordResetEmail($this));
+    }
+
+    public function getFullNameAttribute(): string
+    {
+        return $this->first_name . ' ' . $this->last_name;
+    }
+
+    public function getProfileCoverAttribute()
+    {
+        if (!$this->avatar) {
+            return url('assets/img/profile.png');
+        }
+
+        if (str_contains($this->avatar, "http")) {
+            return $this->avatar;
+        }
+
+        return getFileCDN($this->avatar);
+    }
 }
