@@ -1,6 +1,6 @@
 <?php
 
-return [
+return array(
 
     /*
     |--------------------------------------------------------------------------
@@ -17,4 +17,4 @@ return [
 
     'driver' => 'gd'
 
-];
+);
