@@ -1,47 +1,54 @@
-@extends('master')
+@extends('layouts.auth')
+
+@section('page-title', __('Reset Password'))
 
 @section('content')
-<div class="container">
-    <div class="row">
-        <div class="col-md-8 col-md-offset-2">
-            <div class="panel panel-default">
-                <div class="panel-heading">Reset Password</div>
 
-                <div class="panel-body">
-                    @if (session('status'))
-                        <div class="alert alert-success">
-                            {{ session('status') }}
-                        </div>
-                    @endif
+    <div class="col-md-8 col-lg-6 col-xl-5 mx-auto my-10p">
+        <div class="text-center">
+            <img src="{{ url('assets/img/vanguard-logo.png') }}" alt="{{ setting('app_name') }}" height="50">
+        </div>
 
-                    <form class="form-horizontal" method="POST" action="{{ route('password.email') }}">
+        <div class="card mt-5">
+            <div class="card-body">
+                <h5 class="card-title text-center mt-4 mb-2 text-uppercase">
+                    @lang('Forgot Your Password?')
+                </h5>
+
+                <div class="p-4">
+                    <form role="form" action="<?= route('password.email') ?>" method="POST" id="remind-password-form"
+                          autocomplete="off">
                         {{ csrf_field() }}
 
-                        <div class="form-group{{ $errors->has('email') ? ' has-error' : '' }}">
-                            <label for="email" class="col-md-4 control-label">E-Mail Address</label>
+                        <p class="text-muted mb-4 text-center font-weight-light">
+                            @lang('Please provide your email below and we will send you a password reset link.')
+                        </p>
 
-                            <div class="col-md-6">
-                                <input id="email" type="email" class="form-control" name="email" value="{{ old('email') }}" required>
+                        @include('partials.messages')
 
-                                @if ($errors->has('email'))
-                                    <span class="help-block">
-                                        <strong>{{ $errors->first('email') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
+                        <div class="form-group password-field my-3">
+                            <label for="password" class="sr-only">@lang('Email')</label>
+                            <input type="email"
+                                   name="email"
+                                   id="email"
+                                   class="form-control input-solid"
+                                   placeholder="@lang('Your E-Mail')">
                         </div>
 
-                        <div class="form-group">
-                            <div class="col-md-6 col-md-offset-4">
-                                <button type="submit" class="btn btn-primary">
-                                    Send Password Reset Link
-                                </button>
-                            </div>
+                        <div class="form-group mt-4">
+                            <button type="submit" class="btn btn-primary btn-lg btn-block" id="btn-reset-password">
+                                @lang('Reset Password')
+                            </button>
                         </div>
                     </form>
                 </div>
             </div>
         </div>
+
     </div>
-</div>
-@endsection
+
+@stop
+
+@section('scripts')
+    {!! JsValidator::formRequest('Vanguard\Http\Requests\Auth\PasswordRemindRequest', '#remind-password-form') !!}
+@stop
