@@ -1,70 +1,71 @@
-@extends('master')
+@extends('layouts.auth')
+
+@section('page-title', __('Reset Password'))
 
 @section('content')
-<div class="container">
-    <div class="row">
-        <div class="col-md-8 col-md-offset-2">
-            <div class="panel panel-default">
-                <div class="panel-heading">Reset Password</div>
 
-                <div class="panel-body">
-                    <form class="form-horizontal" method="POST" action="{{ route('password.request') }}">
-                        {{ csrf_field() }}
+    <div class="col-md-8 col-lg-6 col-xl-5 mx-auto my-10p">
+        <div class="text-center">
+            <img src="{{ url('assets/img/vanguard-logo.png') }}" alt="{{ setting('app_name') }}" height="50">
+        </div>
 
-                        <input type="hidden" name="token" value="{{ $token }}">
+        <div class="card mt-5">
+            <div class="card-body">
+                <h5 class="card-title text-center mt-4 mb-2 text-uppercase">
+                    @lang('Reset Your Password')
+                </h5>
 
-                        <div class="form-group{{ $errors->has('email') ? ' has-error' : '' }}">
-                            <label for="email" class="col-md-4 control-label">E-Mail Address</label>
+                <form role="form" action="{{ route('password.update') }}" method="POST" id="reset-password-form"
+                      autocomplete="off" class="p-4">
+                    <input type="hidden" name="token" value="{{ $token }}">
+                    {{ csrf_field() }}
 
-                            <div class="col-md-6">
-                                <input id="email" type="email" class="form-control" name="email" value="{{ $email or old('email') }}" required autofocus>
+                    <p class="text-muted mb-4 text-center font-weight-light px-2">
+                        @lang('Please provide your email and pick a new password below.')
+                    </p>
 
-                                @if ($errors->has('email'))
-                                    <span class="help-block">
-                                        <strong>{{ $errors->first('email') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
-                        </div>
+                    @include('partials.messages')
 
-                        <div class="form-group{{ $errors->has('password') ? ' has-error' : '' }}">
-                            <label for="password" class="col-md-4 control-label">Password</label>
+                    <div class="form-group">
+                        <label for="password" class="sr-only">@lang('Your E-Mail')</label>
+                        <input type="email"
+                               name="email"
+                               id="email"
+                               class="form-control input-solid"
+                               placeholder="@lang('Your E-Mail')">
+                    </div>
 
-                            <div class="col-md-6">
-                                <input id="password" type="password" class="form-control" name="password" required>
+                    <div class="form-group">
+                        <label for="password" class="sr-only">@lang('New Password')</label>
+                        <input type="password"
+                               name="password"
+                               id="password"
+                               class="form-control input-solid"
+                               placeholder="@lang('New Password')">
+                    </div>
 
-                                @if ($errors->has('password'))
-                                    <span class="help-block">
-                                        <strong>{{ $errors->first('password') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
-                        </div>
+                    <div class="form-group">
+                        <label for="password" class="sr-only">@lang('Confirm New Password')</label>
+                        <input type="password"
+                               name="password_confirmation"
+                               id="password_confirmation"
+                               class="form-control input-solid"
+                               placeholder="@lang('Confirm New Password')">
+                    </div>
 
-                        <div class="form-group{{ $errors->has('password_confirmation') ? ' has-error' : '' }}">
-                            <label for="password-confirm" class="col-md-4 control-label">Confirm Password</label>
-                            <div class="col-md-6">
-                                <input id="password-confirm" type="password" class="form-control" name="password_confirmation" required>
+                    <div class="form-group mt-5">
+                        <button type="submit" class="btn btn-primary btn-lg btn-block" id="btn-reset-password">
+                            @lang('Update Password')
+                        </button>
+                    </div>
 
-                                @if ($errors->has('password_confirmation'))
-                                    <span class="help-block">
-                                        <strong>{{ $errors->first('password_confirmation') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
-                        </div>
-
-                        <div class="form-group">
-                            <div class="col-md-6 col-md-offset-4">
-                                <button type="submit" class="btn btn-primary">
-                                    Reset Password
-                                </button>
-                            </div>
-                        </div>
-                    </form>
-                </div>
+                </form>
             </div>
         </div>
     </div>
-</div>
-@endsection
+
+@stop
+
+@section('scripts')
+    {!! JsValidator::formRequest('Vanguard\Http\Requests\Auth\PasswordResetRequest', '#reset-password-form') !!}
+@stop
