@@ -1,77 +1,125 @@
-@extends('master')
+@extends('layouts.auth')
+
+@section('page-title', __('Sign Up'))
+
+@if (setting('registration.captcha.enabled'))
+    <script src='https://www.google.com/recaptcha/api.js'></script>
+@endif
 
 @section('content')
-<div class="container">
-    <div class="row">
-        <div class="col-md-8 col-md-offset-2">
-            <div class="panel panel-default">
-                <div class="panel-heading">Register</div>
 
-                <div class="panel-body">
-                    <form class="form-horizontal" method="POST" action="{{ route('register') }}">
-                        {{ csrf_field() }}
+    <div class="col-md-8 col-lg-6 col-xl-5 mx-auto my-10p">
+        <div class="text-center">
+            <img src="{{ url('assets/img/vanguard-logo.png') }}" alt="{{ setting('app_name') }}" height="50">
+        </div>
 
-                        <div class="form-group{{ $errors->has('name') ? ' has-error' : '' }}">
-                            <label for="name" class="col-md-4 control-label">Name</label>
+        <div class="card mt-5">
+            <div class="card-body">
+                <h5 class="card-title text-center mt-4 text-uppercase">
+                    @lang('Register')
+                </h5>
 
-                            <div class="col-md-6">
-                                <input id="name" type="text" class="form-control" name="name" value="{{ old('name') }}" required autofocus>
+                <div class="p-4">
+                    @include('auth.social.buttons')
 
-                                @if ($errors->has('name'))
-                                    <span class="help-block">
-                                        <strong>{{ $errors->first('name') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
-                        </div>
+                    @include('partials/messages')
 
-                        <div class="form-group{{ $errors->has('email') ? ' has-error' : '' }}">
-                            <label for="email" class="col-md-4 control-label">E-Mail Address</label>
-
-                            <div class="col-md-6">
-                                <input id="email" type="email" class="form-control" name="email" value="{{ old('email') }}" required>
-
-                                @if ($errors->has('email'))
-                                    <span class="help-block">
-                                        <strong>{{ $errors->first('email') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
-                        </div>
-
-                        <div class="form-group{{ $errors->has('password') ? ' has-error' : '' }}">
-                            <label for="password" class="col-md-4 control-label">Password</label>
-
-                            <div class="col-md-6">
-                                <input id="password" type="password" class="form-control" name="password" required>
-
-                                @if ($errors->has('password'))
-                                    <span class="help-block">
-                                        <strong>{{ $errors->first('password') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
-                        </div>
-
+                    <form role="form" action="<?= url('register') ?>" method="post" id="registration-form"
+                          autocomplete="off" class="mt-3">
+                        <input type="hidden" value="<?= csrf_token() ?>" name="_token">
                         <div class="form-group">
-                            <label for="password-confirm" class="col-md-4 control-label">Confirm Password</label>
-
-                            <div class="col-md-6">
-                                <input id="password-confirm" type="password" class="form-control" name="password_confirmation" required>
-                            </div>
+                            <input type="email"
+                                   name="email"
+                                   id="email"
+                                   class="form-control input-solid"
+                                   placeholder="@lang('Email')"
+                                   value="{{ old('email') }}">
+                        </div>
+                        <div class="form-group">
+                            <input type="text"
+                                   name="username"
+                                   id="username"
+                                   class="form-control input-solid"
+                                   placeholder="@lang('Username')"
+                                   value="{{ old('username') }}">
+                        </div>
+                        <div class="form-group">
+                            <input type="password"
+                                   name="password"
+                                   id="password"
+                                   class="form-control input-solid"
+                                   placeholder="@lang('Password')">
+                        </div>
+                        <div class="form-group">
+                            <input type="password"
+                                   name="password_confirmation"
+                                   id="password_confirmation"
+                                   class="form-control input-solid"
+                                   placeholder="@lang('Confirm Password')">
                         </div>
 
-                        <div class="form-group">
-                            <div class="col-md-6 col-md-offset-4">
-                                <button type="submit" class="btn btn-primary">
-                                    Register
-                                </button>
+                        @if (setting('tos'))
+                            <div class="custom-control custom-checkbox">
+                                <input type="checkbox" class="custom-control-input" name="tos" id="tos" value="1"/>
+                                <label class="custom-control-label font-weight-normal" for="tos">
+                                    @lang('I accept')
+                                    <a href="#tos-modal" data-toggle="modal">@lang('Terms of Service')</a>
+                                </label>
                             </div>
+                        @endif
+
+                        {{-- Only display captcha if it is enabled --}}
+                        @if (setting('registration.captcha.enabled'))
+                            <div class="form-group my-4">
+                                {!! app('captcha')->display() !!}
+                            </div>
+                        @endif
+                        {{-- end captcha --}}
+
+                        <div class="form-group mt-4">
+                            <button type="submit" class="btn btn-primary btn-lg btn-block" id="btn-login">
+                                @lang('Register')
+                            </button>
                         </div>
                     </form>
                 </div>
             </div>
         </div>
+
+        <div class="text-center text-muted">
+            @if (setting('reg_enabled'))
+                @lang('Already have an account?')
+                <a class="font-weight-bold" href="<?= url("login") ?>">@lang('Login')</a>
+            @endif
+        </div>
+
     </div>
-</div>
-@endsection
+
+    @if (setting('tos'))
+        <div class="modal fade" id="tos-modal" tabindex="-1" role="dialog" aria-labelledby="tos-label">
+            <div class="modal-dialog" role="document">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h5 class="modal-title" id="tos-label">@lang('Terms of Service')</h5>
+                        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                            <span aria-hidden="true">&times;</span>
+                        </button>
+                    </div>
+                    <div class="modal-body">
+                        @include('auth.tos')
+                    </div>
+                    <div class="modal-footer">
+                        <button type="button" class="btn btn-secondary" data-dismiss="modal">
+                            @lang('Close')
+                        </button>
+                    </div>
+                </div>
+            </div>
+        </div>
+    @endif
+
+@stop
+
+@section('scripts')
+    {!! JsValidator::formRequest('Vanguard\Http\Requests\Auth\RegisterRequest', '#registration-form') !!}
+@stop
