@@ -1,36 +1,69 @@
 <?php
 
-use Illuminate\Http\Request;
 
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| is assigned the "api" middleware group. Enjoy building your API!
-|
-*/
 
-Route::middleware('auth:api')->get('/user', function (Request $request) {
-    return $request->user();
+Route::post('login', 'Auth\AuthController@token');
+Route::post('login/social', 'Auth\SocialLoginController@index');
+Route::post('logout', 'Auth\AuthController@logout');
+
+Route::post('register', 'Auth\RegistrationController@index')->middleware('registration');
+
+Route::group(['middleware' => ['guest', 'password-reset']], function () {
+    Route::post('password/remind', 'Auth\Password\RemindController@index');
+    Route::post('password/reset', 'Auth\Password\ResetController@index');
 });
 
-Route::get('marketapp/commodity-prices', 'Api\MarketController@price');
-Route::get('marketapp/category-commodities', 'Api\MarketController@commodity');
-Route::get('marketapp/commodities-categories', 'Api\MarketController@commodities_cat');
-Route::get('marketapp/commodities-list', 'Api\MarketController@commodities_list');
-Route::get('marketapp/latest-products-updates', 'Api\MarketController@latest_product');
-Route::get('marketapp/latest-products-updates-export', 'Api\MarketController@latest_product_export');
-Route::get('marketapp/market-commodity-prices', 'Api\MarketController@market_commodity');
-Route::get('marketapp/markets-list', 'Api\MarketController@markets_list');
-Route::get('marketapp/province-commodity-prices', 'Api\MarketController@province_commodity');
-Route::post('/submit', 'Api\DataController@recieve');
-Route::get('/data', 'Api\DataController@index');
+Route::group(['middleware' => ['auth', 'registration']], function () {
+    Route::post('email/resend', 'Auth\VerificationController@resend');
+    Route::post('email/verify', 'Auth\VerificationController@verify');
+});
+
+Route::group(['middleware' => ['auth', 'verified']], function () {
+    Route::get('me', 'Profile\DetailsController@index');
+    Route::patch('me/details', 'Profile\DetailsController@update');
+    Route::patch('me/details/auth', 'Profile\AuthDetailsController@update');
+    Route::post('me/avatar', 'Profile\AvatarController@update');
+    Route::delete('me/avatar', 'Profile\AvatarController@destroy');
+    Route::put('me/avatar/external', 'Profile\AvatarController@updateExternal');
+    Route::get('me/sessions', 'Profile\SessionsController@index');
+
+    Route::group(['middleware' => 'two-factor'], function () {
+        Route::put('me/2fa', 'Profile\TwoFactorController@update');
+        Route::post('me/2fa/verify', 'Profile\TwoFactorController@verify');
+        Route::delete('me/2fa', 'Profile\TwoFactorController@destroy');
+    });
+
+    Route::get('stats', 'StatsController@index');
+
+    Route::apiResource('users', 'Users\UsersController')->except('show');
+    Route::get('users/{userId}', 'Users\UsersController@show');
+
+    Route::post('users/{user}/avatar', 'Users\AvatarController@update');
+    Route::put('users/{user}/avatar/external', 'Users\AvatarController@updateExternal');
+    Route::delete('users/{user}/avatar', 'Users\AvatarController@destroy');
+
+    Route::group(['middleware' => 'two-factor'], function () {
+        Route::put('users/{user}/2fa', 'Users\TwoFactorController@update');
+        Route::post('users/{user}/2fa/verify', 'Users\TwoFactorController@verify');
+        Route::delete('users/{user}/2fa', 'Users\TwoFactorController@destroy');
+    });
+
+    Route::get('users/{user}/sessions', 'Users\SessionsController@index');
+
+    Route::get('/sessions/{session}', 'SessionsController@show');
+    Route::delete('/sessions/{session}', 'SessionsController@destroy');
+
+    Route::apiResource('roles', 'Authorization\RolesController')->except('show');
+    Route::get('/roles/{roleId}', 'Authorization\RolesController@show');
+
+    Route::get("roles/{role}/permissions", 'Authorization\RolePermissionsController@show');
+    Route::put("roles/{role}/permissions", 'Authorization\RolePermissionsController@update');
+
+    Route::apiResource('permissions', 'Authorization\PermissionsController');
+
+    Route::get('/settings', 'SettingsController@index');
+
+    Route::get('/countries', 'CountriesController@index');
+});
 
 
-//Online Market
-Route::get('online/new','Api\OnlineMarketController@index');
-Route::get('online/new/{id}','Api\OnlineMarketController@detail');
-//Route::get('/getlatestprice', 'Api\OnlineMarketController@getlatestprice');
