@@ -4,7 +4,7 @@
  * Laravel - A PHP Framework For Web Artisans
  *
  * @package  Laravel
- * @author   Taylor Otwell <taylor@laravel.com>
+ * @author   Taylor Otwell <taylorotwell@gmail.com>
  */
 
 $uri = urldecode(
